backend/requirements.docker.txt
# Lightweight requirements for Docker deployment
# Does NOT include Whisper or PyTorch (too memory-intensive for small droplets)
# For full functionality including YouTube transcription, use requirements.txt locally

Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Cors
Flask-Limiter
python-dotenv
PyPDF2
langchain
google-generativeai
numpy
werkzeug
scikit-learn
psycopg[binary]
gunicorn
trafilatura
yt-dlp
youtube-transcript-api
Flask-Migrate
Flask-Mail
sentence-transformers
bleach
groq
authlib
reportlab
